refator: tracking-page entities refactor tracking page entities are refactored
