+++ src/3_Entity/Tracking/usePutTrackingImage.js
@@ -1,0 +1,44 @@
+import searchpointConverter from "../../4_Shared/lib/searchpointConverter";
+import { useFetch } from "../../4_Shared/util/apiUtil";
+import React from "react";
+
+const usePutTrackingImage = () => {
+  const [serverState, request, loading] = useFetch();
+
+  const putTrackingImage = async (trackingData) => {
+    if (trackingData.line.length <= 0) {
+      alert("트래킹 데이터가 없습니다!");
+      return;
+    }
+    const searchpoint = await searchpointConverter(trackingData.center);
+    request("PUT", `/tracking`, {
+      line: trackingData?.line,
+      searchpoint,
+      center: trackingData?.center,
+      zoom: trackingData?.zoom,
+      heading: trackingData?.heading,
+      sharing: trackingData?.sharing,
+      color: trackingData?.color,
+      thickness: trackingData?.thickness,
+      background: trackingData?.background,
+    });
+  };
+
+  React.useEffect(() => {
+    if (!loading && serverState) {
+      switch (serverState.status) {
+        case 403:
+          console.log(serverState.message);
+          break;
+        case 429:
+          alert("요청이 너무 많습니다! 잠시 기다려주세요.");
+          break;
+        default:
+          break;
+      }
+    }
+  }, [loading, serverState]);
+
+  return [putTrackingImage]
+};
+export default usePutTrackingImage;
